Jarvis_code/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
av==15.0.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
eval_type_backport==0.2.2
flatbuffers==25.2.10
frozenlist==1.7.0
fuzzywuzzy==0.18.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-speech==2.33.0
google-cloud-texttospeech==2.27.0
google-genai==1.26.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.71
langchain-text-splitters==0.3.8
langsmith==0.4.8
Levenshtein==0.27.1
livekit==1.0.12
livekit-agents==1.2.1
livekit-api==1.0.3
livekit-blingfire==1.0.0
livekit-plugins-google==1.2.1
livekit-plugins-noise-cancellation==0.2.5
livekit-plugins-openai==1.2.1
livekit-plugins-silero==1.2.1
livekit-protocol==1.0.4
lxml==6.0.0
marshmallow==3.26.1
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.1
onnxruntime==1.22.1
openai==1.97.1
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-exporter-otlp-proto-http==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
orjson==3.11.0
packaging==25.0
pillow==11.3.0
primp==0.15.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pynput==1.8.1
pyperclip==1.9.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytweening==1.2.0
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.41
sympy==1.14.0
tenacity==8.5.0
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
watchfiles==1.1.0
websockets==15.0.1
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0